Service fini pour le moment, + acceuilController

+++ src/main/java/fr/eni/tp/spring_encheres/bll/EnchereService.java
@@ -5,7 +5,7 @@ import fr.eni.tp.spring_encheres.bo.Enchere;
 import java.util.List;
 
 public interface EnchereService {
-    List<Enchere> findByArticleId(Integer idArticle);
+    List<Enchere> findByArticleId(long idArticle);
 
     void enregistrerEnchere(Enchere enchere);
